% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CA.R
\name{ggplotit.CA}
\alias{ggplotit.CA}
\title{Plot the CA objects with ggplot2}
\usage{
\method{ggplotit}{CA}(x, arrows = c(T, T), ...)
}
\arguments{
\item{x}{An object of class CA}

\item{arrows}{If TRUE arrows will be ploted}

\item{...}{Other parameters}
}
\value{
An ggplot2 plot
}
\description{
Function \code{ggplotit.CA} plots objects of the class CA
}
\examples{
\dontrun{
library(FactoMineR)
obj <- CA(children)
ggplotit(obj)
}
}

