% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cuminc.R
\name{ggplotit.Cuminc}
\alias{ggplotit.Cuminc}
\title{Plot the Cuminc Objects with the ggplot2 Package}
\usage{
\method{ggplotit}{Cuminc}(x, conf.int = FALSE, ...)
}
\arguments{
\item{x}{An object of class Cuminc}

\item{...}{Other parameters}
}
\value{
An ggplot2 plot
}
\description{
Function \code{ggplotit.Cuminc} plots objects of the class Cuminc{mstate}
Calculate nonparametric cumulative incidence functions.
}
\examples{
\dontrun{
library(mstate)
data(aidssi)
ci <- Cuminc(time=aidssi$time, status=aidssi$status)
ggplotit(ci)
}
}

