% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cuminc.R
\name{ggplotit.survfitms}
\alias{ggplotit.survfitms}
\title{Plot the Multi State Objects with the ggplot2 Package}
\usage{
\method{ggplotit}{survfitms}(obj, conf.int = FALSE, labels = NULL, ...)
}
\arguments{
\item{conf.int}{- shall the confidence interval be plotted?}

\item{labels}{- shall labels be added?}

\item{...}{Other parameters}

\item{x}{An object of class survfitms}
}
\value{
An ggplot2 plot
}
\description{
Function \code{ggplotit.survfitms} plots objects of the class survfitms{survival}
Calculate nonparametric cumulative incidence functions.
Competing Risks Model.
}
\examples{
\dontrun{
library(mstate)
data(aidssi)
ci <- survfit(Surv(time=aidssi$time, status=aidssi$status)~1)
ggplotit(ci)
}
}

