% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casmall.R
\name{ggplotit.ca}
\alias{ggplotit.ca}
\title{Plot the ca Objects with the ggplot2 Package}
\usage{
\method{ggplotit}{ca}(x, arrows = c(FALSE, FALSE), ...)
}
\arguments{
\item{x}{An object of class ca}

\item{arrows}{If TRUE arrows will be ploted}

\item{...}{Other parameters}
}
\value{
An ggplot2 plot
}
\description{
Function \code{ggplotit.ca} plots objects of the class ca{ca}
}
\examples{
\dontrun{
library(ca)
obj <- ca(haireye)
ggplotit(obj)
}
}

