% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ggplotit.table}
\alias{ggplotit.table}
\title{Plot the table Objects with the ggplot2 Package}
\usage{
\method{ggplotit}{table}(x, type = "seq", ...)
}
\arguments{
\item{x}{An object of class table}

\item{type}{One of seq (sequential), div (diverging) or qual (qualitative)}

\item{...}{Other parameters}
}
\value{
An ggplot2 plot
}
\description{
Function \code{ggplotit.table} plots objects of the class table{base}
}
\examples{
\dontrun{
library(FactoMineR)
ggplotit(as.table(children))
}
}

